Patch for XALANJ-2175 reviewed by Sarah McNamara (mcnamara () ca ! ibm ! com).
Updated JDK version test so that -Xbootclasspath option will be specified for
spawned JVM if the version is not one of 1.1, 1.2 or 1.3.  That allows the
test harness to work on JDK 1.1 through 1.5, and I hope it will allow us to
avoid the need to add checks for more JDK versions in the future.